refactor: make projections use native serialization by default
Not using it will make nested objects not work if they are not of the same type, therefore making deep projection impossible. To allow for deep projection, the data will be converted to a `native` json map. The old behaviour can still be used using the `-Shallow` version of the method or the shallow argument.