feat!: migrate to the new element model and remove SerializableLibrary
SerializableLibrary has been removed to simplify the dogs_generator and make it easier to keep it up to date with the upstream analyzer. Porting this to the new directive-based annotation discovery would have been hell since I would have to rewrite portions of lyell for this to properly work. Additionally, the feature was finicky to use and prone to errors anyway.

Refs: #19